New Features: create schedule can have courses in view
update notice:
-*course schedule(pre-defined in create a new course section) will have
a higher priority in user schedule view and creation. Any user schedule
slots which have overlaps with any of the course schedule slots will be
overridden(removed).
--view/edit schedule page and create schedule page now have the ability
to reference the active course time in the schedule when users are
making adjustment of the schedule or creating a new schedule by looking
at the course schedule layout at the same time.
--duplicate options:
---cannot duplicate or add or remove a course from there, and the
serializing functions will not record the course schedule when submit
---course schedule time slots have special properties other than normal
user schedule slots. Course schedule slots will hinder the insertion of
schedule with overlapping time intervals. Course schedule slots will not
be duplicated in duplicate mode. In addition, course schedule slots will
be counted in duplicate mode validation, and if none of the duplicate
schedule slots conflicts with the course schedule slots in the target
day/week, user schedule slots can be therefore copied.