Bugs fixed:
1. Student register console does not need to upload the certificate
files(put the file upload into the instructor register section in the if
clause).
2.New admitted instructor can pass the creating a new schedule generate
button validation(if the active courses list is empty, give the start
and end date the info that the user provided w/o making any adjustment.
3.Fixed bugs in the final serializing function(only record the start and end
date/time info as ISOString provides.
4.Fixed bus in the serializing function(visuals to strings). The
problems were at the last row of the calendar. Changed the
endTime(automatically set to 00:00) to the time at the previous thead + 15mins.